+++ packages/cmrjs/tests/cmr-utils/test-reconcilecmrmetadata.js
@@ -350,7 +350,7 @@ test('publishUMMGJSON2CMR calls ingestUMMGranule with ummgMetadata via valid CMR
     granuleId: 'fakeGranuleID',
   };
   const publishUMMGJSON2CMR = cmrUtils.__get__('publishUMMGJSON2CMR');
-  const ingestFake = sinon.fake.resolves({ 'concept-id': 'fakeID', 'revision-id': 123 });
+  const ingestFake = sinon.fake.resolves({ 'concept-id': 'fakeID' });
   const CmrFake = sinon.fake.returns({ ingestUMMGranule: ingestFake });
 
   const restoreCMR = cmrUtils.__set__('CMR', CmrFake);